// src/types/api/service/models/NewsAlertDto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * MediaBee Service API
 * MediaBee Service API
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface NewsAlertDto
 */
export interface NewsAlertDto {
    /**
     * 알리미 ID
     * @type {number}
     * @memberof NewsAlertDto
     */
    alertId?: number;
    /**
     * 모니터링 ID
     * @type {number}
     * @memberof NewsAlertDto
     */
    newsSrchId: number;
    /**
     * 뉴스 알리미 제목
     * @type {string}
     * @memberof NewsAlertDto
     */
    title: string;
    /**
     * 뉴스 알리미 내용
     * @type {string}
     * @memberof NewsAlertDto
     */
    content?: string;
    /**
     * 정렬 옵션 (게제일, 관련성, 글자수) - common_field.NEWS_ALERT_SORT
     * @type {string}
     * @memberof NewsAlertDto
     */
    sortOption: string;
}

/**
 * Check if a given object implements the NewsAlertDto interface.
 */
export function instanceOfNewsAlertDto(value: object): value is NewsAlertDto {
    if (!('newsSrchId' in value) || value['newsSrchId'] === undefined) return false;
    if (!('title' in value) || value['title'] === undefined) return false;
    if (!('sortOption' in value) || value['sortOption'] === undefined) return false;
    return true;
}

export function NewsAlertDtoFromJSON(json: any): NewsAlertDto {
    return NewsAlertDtoFromJSONTyped(json, false);
}

export function NewsAlertDtoFromJSONTyped(json: any, ignoreDiscriminator: boolean): NewsAlertDto {
    if (json == null) {
        return json;
    }
    return {
        
        'alertId': json['alertId'] == null ? undefined : json['alertId'],
        'newsSrchId': json['newsSrchId'],
        'title': json['title'],
        'content': json['content'] == null ? undefined : json['content'],
        'sortOption': json['sortOption'],
    };
}

export function NewsAlertDtoToJSON(value?: NewsAlertDto | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'alertId': value['alertId'],
        'newsSrchId': value['newsSrchId'],
        'title': value['title'],
        'content': value['content'],
        'sortOption': value['sortOption'],
    };
}